+++ app/Model/OauthUserToken.php
@@ -6,7 +6,10 @@ use Illuminate\Database\Eloquent\Model;
 
 class OauthUserToken extends Model
 {
-    //
+    protected $connection = 'mysql';
+
+    public static $alias = 'oauth_user_token';
+    
     /**
      * The attributes that are mass assignable.
      *
